Luna: icon pack/app labeling integration
Adds support for changing application icons, as well as changing
the label of the application. You can change all of the application
icons, or choose which apps you want to change.

Editing the application label is simple. Long pressing the application
and dragging it to the "Edit" drop target will bring up the editing menu
to change the application label. You can change the icon by pressing the
current icon and select the icon pack you wish to choose the icon from.
This will then bring up a icon chooser view to select any icon available
from the chosen icon pack.

This is based on Paranoid Android implementation, without the
customization
menu and adapted with our current iconprovider. Full credits go to @Anas
for
the base design.

Change-Id: I081293f6a5457bdd3a9e0cb678fb71a4206ae733