Luna: app shortcuts/notification badges integration
This is a port of a backport from the new Pixel Launcher
with Android O exclusives. This apps the new app shortcuts
design with notification badges from Android O. This also
adds the new widgets view.

Full credits to @deletescape for backporting everything into
lawnchair. I only ported it over to work with Luna/Launcher3

Change-Id: Ie307abc516dd05db9ed2824e46a79fd514b5d5ff