Luna: add preference to enable the notification badge service
Notification badges are disabled by default and require enabling
the service listener in notification access settings. Not all users
know this, so add a preference to launch notification access settings
to turn it on.

Added a boolean function to obtain the current state of the service. If
disabled, show the preference in the settings menu requesting the user
to enable it. If enabled, remove the preference completely.

Change-Id: Iffe35f1d7b2910e79b87c1e3dd0e3ae93afd9e9f