Quickspace: Initial support for QuickEvents
This adds a new UI to Quickspace when events are available such as Now
Playing, Alarms, or whatever we configure as an event. This logic is
inspired by Google's event infrastructure in Smartspace but completely
written from scratch to be housed in Quickspace.

Here we can create our own events and maybe even allow thirdparty events

Change-Id: I96e84c8600b37ad61dcb2cf27d59dca84db490eb